build(run-e2e): splits deploy script into 2 sections for netlify ci
